Remove the {deploy,destroy}-rds-broker errand jobs.
These are used to deploy the rds-broker as a CloudFoundry application
instead of running it on a dedicated VM. This is not something we need
to do becuase it makes it hard/impossible to use IAM instance profiles
to provide credentials to the broker.

Instead we're deploying the broker to its own VM with the necessary
instance profile.