Use our own release blobs bucket
We want to upgrade the RDS broker Golang version to 1.8 to support more
recent libraries for the brokerapi and aws-sdk-go. The 3rd party bucket
we were using did not have the blobs to support this, so we will have to
use our own.

This references a bucket created by our CI build environment for Bosh
releases. The region and endpoint configurations are required so Bosh
doesn't assume it is us-east-1.