Remove route-registrar code and configuration
This is no longer used. It existed for situations when the broker
is deployed in a way that requires its route to be registered with
the Cloud Foundry router. We have a dedicated ELB in front of the broker,
so the route-registrar is now redundant.

The route-registrar code was a submodule, so I followed well-documented
steps[1] for its removal, as there is no all-mighty solution in Git for
this. The commands run (in addition to removal of job and package files):

```
git submodule deinit src/github.com/cloudfoundry-incubator/route-registrar
git rm src/github.com/cloudfoundry-incubator/route-registrar
```

I left the contents of `releases/aws-broker/*`, because it is still true
that these releases were once built.

[1] http://stackoverflow.com/a/16162000/4470895