Remove .final_builds directory
None of the files in here relate to the releases we've been creating,
they predate us forking this from the origin upstream POC repo. We're
not currently creating final releases, we're using dev releases
everywhere, and having Concourse upload them to the releases bucket.

Removing them because they're only misleading, and not useful.