Pass certs and keys directly as strings instead of paths
It's a bad practice to refer directly to files in other releases since
you might end up with releases in different containers which don't have
access to each others file systems.

Instead we can pass the content of the file directly and create the
certs and keys with trivial templates. We can then hardcode the paths of
these files in config.json.